requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.5
colorlog==6.7.0
frozenlist==1.3.1
hikari==2.0.0.dev115
hikari-lightbulb==2.2.4
idna==3.3
multidict==6.0.2
python-dotenv==0.20.0
requests==2.28.1
urllib3==1.26.12
yarl==1.8.1
statbotics==2.0.1